# sdnnt/src/main/java/cz/inovatika/sdnnt/openapi/endpoints/model/CatalogObject.java
/*
 * SDNNT API
 *  API umožnuje vzdáleně spravovat svoje žádosti, vytvářet nové, prohlížet již poslané a procházet katalog.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package cz.inovatika.sdnnt.openapi.endpoints.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * CatalogObject
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-07-15T08:56:28.035Z[GMT]")public class CatalogObject   {
  @JsonProperty("batch")
  private CatalogObjectBatch batch = null;

  public CatalogObject batch(CatalogObjectBatch batch) {
    this.batch = batch;
    return this;
  }

  /**
   * Get batch
   * @return batch
   **/
  @JsonProperty("batch")
  @Schema(description = "")
  @Valid
  public CatalogObjectBatch getBatch() {
    return batch;
  }

  public void setBatch(CatalogObjectBatch batch) {
    this.batch = batch;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    CatalogObject catalogObject = (CatalogObject) o;
    return Objects.equals(this.batch, catalogObject.batch);
  }

  @Override
  public int hashCode() {
    return Objects.hash(batch);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class CatalogObject {\n");
    
    sb.append("    batch: ").append(toIndentedString(batch)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}