# sdnnt/src/main/java/cz/inovatika/sdnnt/openapi/endpoints/model/NotSavedRequest.java
/*
 * SDNNT API
 *  API umožnuje vzdáleně spravovat svoje žádosti, vytvářet nové, prohlížet již poslané a procházet katalog.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package cz.inovatika.sdnnt.openapi.endpoints.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;

import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * NotSavedRequest
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-07-15T08:56:28.035Z[GMT]")public class NotSavedRequest extends Request  {
  @JsonProperty("id")
  private String id = null;

  @JsonProperty("navrh")
  private String navrh = null;

  @JsonProperty("user")
  private String user = null;

  @JsonProperty("state")
  private String state = null;

  @JsonProperty("reason")
  private String reason = null;

  public NotSavedRequest id(String id) {
    this.id = id;
    return this;
  }

  /**
   * Get id
   * @return id
   **/
  @JsonProperty("id")
  @Schema(description = "")
  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }

  public NotSavedRequest navrh(String navrh) {
    this.navrh = navrh;
    return this;
  }

  /**
   * Get navrh
   * @return navrh
   **/
  @JsonProperty("navrh")
  @Schema(description = "")
  public String getNavrh() {
    return navrh;
  }

  public void setNavrh(String navrh) {
    this.navrh = navrh;
  }

  public NotSavedRequest user(String user) {
    this.user = user;
    return this;
  }

  /**
   * Get user
   * @return user
   **/
  @JsonProperty("user")
  @Schema(description = "")
  public String getUser() {
    return user;
  }

  public void setUser(String user) {
    this.user = user;
  }

  public NotSavedRequest state(String state) {
    this.state = state;
    return this;
  }

  /**
   * Get state
   * @return state
   **/
  @JsonProperty("state")
  @Schema(description = "")
  public String getState() {
    return state;
  }

  public void setState(String state) {
    this.state = state;
  }

  public NotSavedRequest reason(String reason) {
    this.reason = reason;
    return this;
  }

  /**
   * Get reason
   * @return reason
   **/
  @JsonProperty("reason")
  @Schema(description = "")
  public String getReason() {
    return reason;
  }

  public void setReason(String reason) {
    this.reason = reason;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    NotSavedRequest notSavedRequest = (NotSavedRequest) o;
    return Objects.equals(this.id, notSavedRequest.id) &&
        Objects.equals(this.navrh, notSavedRequest.navrh) &&
        Objects.equals(this.user, notSavedRequest.user) &&
        Objects.equals(this.state, notSavedRequest.state) &&
        Objects.equals(this.reason, notSavedRequest.reason) &&
        super.equals(o);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, navrh, user, state, reason, super.hashCode());
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class NotSavedRequest {\n");
    sb.append("    ").append(toIndentedString(super.toString())).append("\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    navrh: ").append(toIndentedString(navrh)).append("\n");
    sb.append("    user: ").append(toIndentedString(user)).append("\n");
    sb.append("    state: ").append(toIndentedString(state)).append("\n");
    sb.append("    reason: ").append(toIndentedString(reason)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}