# sdnnt/src/main/java/cz/inovatika/sdnnt/openapi/endpoints/model/Request.java
/*
 * SDNNT API
 *  API umožnuje vzdáleně spravovat svoje žádosti, vytvářet nové, prohlížet již poslané a procházet katalog.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package cz.inovatika.sdnnt.openapi.endpoints.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import java.util.ArrayList;
import java.util.Date;
import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * Request
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-07-15T08:56:28.035Z[GMT]")public class Request   {
  @JsonProperty("identifiers")
  private List<String> identifiers = null;

  @JsonProperty("datum_zadani")
  private Date datumZadani = null;

  @JsonProperty("pozadavek")
  private String pozadavek = null;

  @JsonProperty("poznamka")
  private String poznamka = null;

  public Request identifiers(List<String> identifiers) {
    this.identifiers = identifiers;
    return this;
  }

  public Request addIdentifiersItem(String identifiersItem) {
    if (this.identifiers == null) {
      this.identifiers = new ArrayList<String>();
    }
    this.identifiers.add(identifiersItem);
    return this;
  }

  /**
   * Get identifiers
   * @return identifiers
   **/
  @JsonProperty("identifiers")
  @Schema(description = "")
  public List<String> getIdentifiers() {
    return identifiers;
  }

  public void setIdentifiers(List<String> identifiers) {
    this.identifiers = identifiers;
  }

  public Request datumZadani(Date datumZadani) {
    this.datumZadani = datumZadani;
    return this;
  }

  /**
   * Get datumZadani
   * @return datumZadani
   **/
  @JsonProperty("datum_zadani")
  @Schema(description = "")
  @Valid
  public Date getDatumZadani() {
    return datumZadani;
  }

  public void setDatumZadani(Date datumZadani) {
    this.datumZadani = datumZadani;
  }

  public Request pozadavek(String pozadavek) {
    this.pozadavek = pozadavek;
    return this;
  }

  /**
   * Get pozadavek
   * @return pozadavek
   **/
  @JsonProperty("pozadavek")
  @Schema(description = "")
  public String getPozadavek() {
    return pozadavek;
  }

  public void setPozadavek(String pozadavek) {
    this.pozadavek = pozadavek;
  }

  public Request poznamka(String poznamka) {
    this.poznamka = poznamka;
    return this;
  }

  /**
   * Get poznamka
   * @return poznamka
   **/
  @JsonProperty("poznamka")
  @Schema(description = "")
  public String getPoznamka() {
    return poznamka;
  }

  public void setPoznamka(String poznamka) {
    this.poznamka = poznamka;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Request request = (Request) o;
    return Objects.equals(this.identifiers, request.identifiers) &&
        Objects.equals(this.datumZadani, request.datumZadani) &&
        Objects.equals(this.pozadavek, request.pozadavek) &&
        Objects.equals(this.poznamka, request.poznamka);
  }

  @Override
  public int hashCode() {
    return Objects.hash(identifiers, datumZadani, pozadavek, poznamka);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Request {\n");
    
    sb.append("    identifiers: ").append(toIndentedString(identifiers)).append("\n");
    sb.append("    datumZadani: ").append(toIndentedString(datumZadani)).append("\n");
    sb.append("    pozadavek: ").append(toIndentedString(pozadavek)).append("\n");
    sb.append("    poznamka: ").append(toIndentedString(poznamka)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}